ajuste telas cadastroCargo e cadastroQuarto e ajustes nos selects

+++ hotel/src/main/java/br/com/hotel/entidades/Cargo.java
@@ -13,16 +13,16 @@ public class Cargo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(nullable = false)
-    private int idCargo;
+    private Long idCargo;
     @Column(nullable = false)
     private String tipoCargo;
     @Column(nullable = false)
     private Boolean permissao;
 
-    public int getIdCargo() {
+    public Long getIdCargo() {
         return idCargo;
     }
-    public void setIdCargo(int idCargo) {
+    public void setIdCargo(Long idCargo) {
         this.idCargo = idCargo;
     }
     public String getTipoCargo() {
